+++ internal/cmd/default.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/AdguardTeam/AdGuardDNSClient/internal/agdc"
+	"github.com/AdguardTeam/AdGuardDNSClient/internal/configmigrate"
 	"github.com/AdguardTeam/golibs/errors"
 	"github.com/AdguardTeam/golibs/logutil/slogutil"
 	"github.com/AdguardTeam/golibs/netutil"
@@ -16,7 +17,16 @@ import (
 	"gopkg.in/yaml.v3"
 )
 
+// Values for the default DNS configuration.
 const (
+	// defaultUpstreamAddress is the address of common upstream DNS server to
+	// use by default.
+	defaultUpstreamAddress = "https://unfiltered.adguard-dns.com/dns-query"
+
+	// defaultFallbackAddress is the address of the fallback upstream DNS server
+	// to use by default.
+	defaultFallbackAddress = "tls://94.140.14.140"
+
 	// defaultPlainDNSPort is the default port for plain DNS.
 	defaultPlainDNSPort uint16 = 53
 
@@ -24,9 +34,66 @@ const (
 	defaultUpstreamTimeout = 2 * time.Second
 )
 
+// Values for the default server configuration.
+const (
+	// defaultBindRetryEnabled is the default value for the bind retrying
+	// feature to be enabled.
+	//
+	// See https://github.com/AdguardTeam/AdGuardDNSClient/issues/11.
+	defaultBindRetryEnabled = true
+
+	// defaultBindRetryIvl is the default interval to wait between listen
+	// addresses binding retries.
+	defaultBindRetryIvl = 1 * time.Second
+
+	// defaultBindRetryCount is the default maximum number of attempts to bind a
+	// listen address after the first one.
+	defaultBindRetryCount uint = 4
+)
+
+// Values for the default cache configuration.
+const (
+	// defaultCacheEnabled is the default value for the cache usage.
+	defaultCacheEnabled = true
+
+	// defaultCacheSize is the default size of the cache.
+	defaultCacheSize = 128 * datasize.MB
+
+	// defaultCacheClientSize is the default size of the cache for the client.
+	defaultCacheClientSize = 4 * datasize.MB
+)
+
+// Values for the default profiling configuration.
+const (
+	// defaultPprofEnabled is the default value for the pprof server to be
+	// enabled.
+	defaultPprofEnabled = false
+
+	// defaultPprofPort is the default port to serve pprof handlers locally.
+	defaultPprofPort uint16 = 6060
+)
+
+// Values for the default log configuration.
+const (
+	// defaultLogOutput is the default output for the logs.
+	defaultLogOutput = outputSyslog
+
+	// defaultLogFormat is the default format for the logs.
+	defaultLogFormat = slogutil.FormatDefault
+
+	// defaultLogTimestamp is the default value for the timestamp presence in
+	// logs.
+	defaultLogTimestamp = false
+
+	// defaultLogVerbose is the default value for the verbosity of the logs.
+	defaultLogVerbose = false
+)
+
 // filterInterfaceAddrs gets the addresses as given by [net.InterfaceAddrs] and
 // filters out the ones that are not in the set.  It returns the [ipPortConfig]s
 // for the eligible addresses created using port p.
+//
+// TODO(e.burkov):  Use logger instead of [fmt.Fprintf].
 func filterInterfaceAddrs(
 	addrs []net.Addr,
 	set netutil.SubnetSet,
@@ -90,12 +157,17 @@ func newDefaultServerConfig() (c *serverConfig, err error) {
 
 	return &serverConfig{
 		ListenAddresses: localAddrs,
+		BindRetry: &bindRetryConfig{
+			Enabled:  defaultBindRetryEnabled,
+			Interval: timeutil.Duration(defaultBindRetryIvl),
+			Count:    defaultBindRetryCount,
+		},
 	}, nil
 }
 
 // newDefaultDNSConfig creates a new default configuration for DNS.
 func newDefaultDNSConfig() (c *dnsConfig, err error) {
-	defer func() { err = errors.Annotate(err, "creating default DNS configuration: %w") }()
+	defer func() { err = errors.Annotate(err, "creating default dns configuration: %w") }()
 
 	serverConf, err := newDefaultServerConfig()
 	if err != nil {
@@ -114,22 +186,22 @@ func newDefaultDNSConfig() (c *dnsConfig, err error) {
 	}}
 	upstreamGroups := upstreamGroupsConfig{
 		agdc.UpstreamGroupNameDefault: &upstreamGroupConfig{
-			Address: "https://unfiltered.adguard-dns.com/dns-query",
+			Address: defaultUpstreamAddress,
 			// TODO(e.burkov):  It marshals into an empty slice, but should not
 			// appear in the configuration file at all.
 			Match: nil,
 		},
 	}
 	fallbackServers := []*urlConfig{{
-		Address: "tls://94.140.14.140",
+		Address: defaultFallbackAddress,
 	}}
 
 	return &dnsConfig{
 		Server: serverConf,
 		Cache: &cacheConfig{
-			Enabled:    true,
-			Size:       128 * datasize.MB,
-			ClientSize: 4 * datasize.MB,
+			Enabled:    defaultCacheEnabled,
+			Size:       defaultCacheSize,
+			ClientSize: defaultCacheClientSize,
 		},
 		Bootstrap: &bootstrapConfig{
 			Servers: bootstrapServers,
@@ -161,17 +233,17 @@ func newDefaultConfig() (c *configuration, err error) {
 		DNS: dnsConf,
 		Debug: &debugConfig{
 			Pprof: &pprofConfig{
-				Enabled: false,
-				Port:    6060,
+				Enabled: defaultPprofEnabled,
+				Port:    defaultPprofPort,
 			},
 		},
 		Log: &logConfig{
-			Output:    outputSyslog,
-			Format:    slogutil.FormatDefault,
-			Timestamp: false,
-			Verbose:   false,
+			Output:    defaultLogOutput,
+			Format:    defaultLogFormat,
+			Timestamp: defaultLogTimestamp,
+			Verbose:   defaultLogVerbose,
 		},
-		SchemaVersion: currentSchemaVersion,
+		SchemaVersion: configmigrate.VersionLatest,
 	}, nil
 }